+++ backend/routers/integrations.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, Query, HTTPException
+from fastapi import APIRouter, Depends, Query, HTTPException, status
 from enums import UserAuthorizationStatus
 from dependencies import get_integration_service, IntegrationService, IntegrationsPresistence, get_integration_service, get_user_integrations_presistence, \
     check_user_authorization, check_domain, check_pixel_install_domain, check_user_authentication
@@ -31,6 +31,11 @@ async def get_credential_service(platform: str,
 async def export(export_query: ExportLeads, service_name: str = Query(...),
                  integrations_service: IntegrationService = Depends(get_integration_service),
                  user = Depends(check_user_authorization), domain = Depends(check_pixel_install_domain)):
+    if user.team_owner_id is not None and (user.access_level != 'admin' or user.access_level != 'standard'):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Access denied. Admins and standard only."
+        )
     with integrations_service as service: 
         service = getattr(service, service_name)
         service.export_leads(export_query.list_name, user['id'])
@@ -41,6 +46,11 @@ async def export(export_query: ExportLeads, service_name: str = Query(...),
 async def create_integration(creditional: IntegrationCredentials, service_name: str = Query(...), 
                              integration_service: IntegrationService = Depends(get_integration_service),
                              user = Depends(check_user_authentication), domain = Depends(check_domain)):
+    if user.team_owner_id is not None and (user.access_level != 'admin' or user.access_level != 'standard'):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Access denied. Admins and standard only."
+        )
     if not creditional.pixel_install and not domain.is_pixel_installed:
         raise HTTPException(status_code=403, detail={'status': UserAuthorizationStatus.PIXEL_INSTALLATION_NEEDED.value})
     with integration_service as service:
@@ -55,6 +65,11 @@ async def create_integration(creditional: IntegrationCredentials, service_name: 
 async def delete_integration(service_name: str = Query(...),
                              user = Depends(check_user_authorization),
                              integration_service: IntegrationService = Depends(get_integration_service), domain = Depends(check_pixel_install_domain)):
+    if user.team_owner_id is not None and (user.access_level != 'admin' or user.access_level != 'standard'):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Access denied. Admins and standard only."
+        )
     try:
         integration_service.delete_integration(service_name, user)
         return {'message': 'Successfuly'}
@@ -80,6 +95,11 @@ async def get_list(service_name: str = Query(...),
 async def create_sync(data: SyncCreate, service_name: str = Query(...),
                       integration_service: IntegrationService = Depends(get_integration_service),
                       user = Depends(check_user_authorization), domain = Depends(check_pixel_install_domain)):
+    if user.team_owner_id is not None and (user.access_level != 'admin' or user.access_level != 'standard'):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Access denied. Admins and standard only."
+        )
     with integration_service as service:
         service = getattr(service, service_name)
         service.create_sync(user['id'], **data.model_dump())
